backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==24.4.2
automat==25.4.16
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==47.0.0
daphne==4.2.1
django==4.2.30
django-channels==0.7.0
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.1
hyperlink==21.0.0
idna==3.13
incremental==24.11.0
kombu==5.6.2
oauthlib==3.3.1
packaging==26.2
prompt-toolkit==3.0.52
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==2.23
pyopenssl==26.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.1
twisted==25.5.0
txaio==23.6.1
typing-extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope-interface==8.0.1
